requirements.txt
alembic==1.7.5
aniso8601==9.0.1
attrs==21.4.0
click==8.0.3
Flask==2.0.2
Flask-HTTPAuth==4.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
mirakuru==2.4.1
packaging==21.3
pkg_resources==0.0.0
pluggy==1.0.0
port-for==0.6.1
psutil==5.9.0
psycopg==3.0.7
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytest-flask==1.2.0
pytest-postgresql==4.1.0
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.29
toml==0.10.2
Werkzeug==2.0.2